Common/Threading: Move a bunch of unnecessary wrappers to gui
