+++ common/common.vcxproj
@@ -98,9 +98,7 @@
     <ClCompile Include="Windows\WinMisc.cpp" />
     <ClCompile Include="Windows\WinThreads.cpp" />
     <ClCompile Include="Misc.cpp" />
-    <ClCompile Include="Mutex.cpp" />
     <ClCompile Include="Semaphore.cpp" />
-    <ClCompile Include="ThreadTools.cpp" />
     <ClCompile Include="emitter\bmi.cpp" />
     <ClCompile Include="emitter\cpudetect.cpp" />
     <ClCompile Include="emitter\fpu.cpp" />
