+++ common/emitter/legacy.cpp
@@ -26,6 +26,7 @@
 //------------------------------------------------------------------
 
 #include "common/emitter/legacy_internal.h"
+#include "common/Console.h"
 
 emitterT void ModRM(uint mod, uint reg, uint rm)
 {
