+++ pcsx2/gui/Mutex.cpp
@@ -13,7 +13,7 @@
  *  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "common/Threading.h"
+#include "PersistentThread.h"
 
 namespace Threading
 {
